Introduce CompletionChunkKind class
We build the CompletionChunkKinds directly from an enumeration, which allows
us to easily add now CompletionChunkKinds if they are added to libclang.

With the addition of the CompletionChunkKind, we introduce a
spelling cache that avoids function calls to libclang, in case the
spelling already known. We see a good speedup with this, in case there
are a lot of completion results (e.g. 'std::^').

We also use this opportunity to avoid string comparisons when checking
the type kind.