libclang: Specialized clang_complete snippet engine
Instead of post processing the formatted results to add snippet
annotations, we add the snippet annotations during formatting of the
completion results. In case this is done in a specialized formatting
function this does not add any overhead during formatting and it removes
the entire snippet formatting overhead later on.

This reduces code completion time for examples/boost.cpp from 132 to 99
miliseconds.